requirements.txt
folium(Створення мапи, шарів)
geopy.decoders(ARCgis)(отримання координат)